Deprecate all remaining dt_widget_* api
The rationale is that modern HTML5 provides already consistent and localized UI elements for date and time input. It is also difficult to style a widget that comes pre-baked by the api. One should either use acs-templating or xowiki feature to generate UIs, or modern HTML5 is in most cases sufficient.